requirements.txt
Flask==3.0.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
werkzeug==3.0.1
